libpower: Remove device-specific compilation for I2C power HAL
Using target-specific code breaks builds for derived targets (which
will still include the mixin, but which will have a different
$(TARGET_PRODUCT).

This array just stores a list of devices which should be powered down
affirmatively by the power HAL on entering non-interacive mode.
There's no downside to having devices listed that aren't present on
the system -- they'll simply be ignored.

If this list reaches dozens of devices and we have pollution problems,
we might consider making it configurable in the future (though my vote
would be for a runtime list instead of stuffing it into C++ code).
But right now it's one Atmel touchpanel used on one (potentially two
if the derived target in question uses the same panel) device.  I
don't see the point.

Issue: ADEV-556
Change-Id: Id846bb3dd01284c3ea184d816fa5335e81bb09a8
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
Reviewed-on: https://android.intel.com/163049
Reviewed-by: Leung, Daniel <daniel.leung@intel.com>
Tested-by: Leung, Daniel <daniel.leung@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>